+++ app/layouts/admin.vue
@@ -1,56 +1,137 @@
-<!-- layouts/admin.vue -->
 <script setup lang="ts">
-import { authClient } from '~/utils/auth-client'
+import './admin.css'
 
-const route = useRoute()
+const router = useRouter()
 
-async function logout() {
-  const confirmed = confirm('Are you sure you want to log out?')
-  if (!confirmed) return
+const contexts = [
+  {
+    label: 'Admin',
+    value: 'admin',
+    route: '/admin',
+  },
+  {
+    label: '+ Create New Class',
+    value: 'create-class',
+    route: '/coach/create-class',
+  },
+]
 
-  try {
-    await authClient.signOut()
-    window.location.href = '/auth'
-  } catch (error) {
-    console.error('Logout failed:', error)
+const selectedContext = ref('admin')
+
+const navigation = [
+  {
+    label: 'Dashboard',
+    route: '/admin',
+  },
+  {
+    label: 'Data',
+    route: '/admin/data',
+  },
+  {
+    label: 'Teacher Verification',
+    route: '/admin/teacher-verification',
+  },
+]
+
+async function changeContext() {
+  const selected = contexts.find(
+    context => context.value === selectedContext.value,
+  )
+
+  if (!selected) {
+    return
   }
+
+  await router.push(selected.route)
 }
 </script>
 
 <template>
-  <div class="rh-admin-wrap">
-    <aside class="rh-sidebar">
-      <div class="rh-sidebar-inner">
-        <div class="rh-logo">
-          <div class="rh-logo-icon">L</div>
-          <div>
-            <div class="rh-logo-name">Reading<span class="rh-logo-accent">Huddle</span></div>
-            <div class="rh-logo-sub">Admin Portal</div>
+  <div class="portal-layout">
+    <aside class="portal-sidebar">
+      <div class="portal-sidebar-content">
+        <NuxtLink
+          to="/admin"
+          class="portal-logo"
+        >
+          <div class="portal-logo-icon">
+            L
           </div>
-        </div>
 
-        <p class="rh-nav-label">Admin Tools</p>
-        <nav class="rh-nav">
-          <button @click="navigateTo('/admin')"               class="rh-nav-btn" :class="{ 'rh-nav-active': route.path === '/admin' }">Dashboard</button>
-          <button @click="navigateTo('/admin/builder')"       class="rh-nav-btn" :class="{ 'rh-nav-active': route.path === '/admin/builder' }">Form Builder</button>
-          <button @click="navigateTo('/admin/progress')"      class="rh-nav-btn" :class="{ 'rh-nav-active': route.path === '/admin/progress' }">Class Progress</button>
-          <button @click="navigateTo('/admin/raffle')"        class="rh-nav-btn" :class="{ 'rh-nav-active': route.path === '/admin/raffle' }">Raffle System</button>
-          <button @click="navigateTo('/admin/announcements')" class="rh-nav-btn" :class="{ 'rh-nav-active': route.path === '/admin/announcements' }">Announcements</button>
+          <div>
+            <div class="portal-logo-title">
+              Reading<span>Huddle</span>
+            </div>
+
+            <div class="portal-logo-subtitle">
+              Admin Portal
+            </div>
+          </div>
+        </NuxtLink>
+
+        <p class="portal-nav-label">
+          Navigation
+        </p>
+
+        <nav class="portal-navigation">
+          <NuxtLink
+            v-for="item in navigation"
+            :key="item.route"
+            :to="item.route"
+            class="portal-nav-link"
+            exact-active-class="portal-nav-link-active"
+          >
+            {{ item.label }}
+          </NuxtLink>
         </nav>
       </div>
 
-      <div class="rh-sidebar-footer">
-        <NuxtLink to="/" class="rh-back-link">← Back to Portal</NuxtLink>
-        <button class="rh-logout-link" @click="logout">⎋ Logout</button>
+      <div class="portal-sidebar-footer">
+        <NuxtLink
+          to="/auth"
+          class="portal-logout"
+        >
+          ← Log out
+        </NuxtLink>
       </div>
     </aside>
 
-    <div class="rh-main">
-      <slot />
+    <div class="portal-workspace">
+      <header class="portal-header">
+        <div>
+          <p class="portal-header-label">
+            Reading Huddle
+          </p>
+
+          <h1 class="portal-header-title">
+            Administration
+          </h1>
+        </div>
+
+        <div class="portal-context">
+          <label for="portal-context">
+            Managing
+          </label>
+
+          <select
+            id="portal-context"
+            v-model="selectedContext"
+            @change="changeContext"
+          >
+            <option
+              v-for="context in contexts"
+              :key="context.value"
+              :value="context.value"
+            >
+              {{ context.label }}
+            </option>
+          </select>
+        </div>
+      </header>
+
+      <main class="portal-main">
+        <slot />
+      </main>
     </div>
   </div>
 </template>
-
-<style>
-@import './admin.css';
-</style>
