+++ app/middleware/auth.global.ts
@@ -7,9 +7,10 @@ export default defineNuxtRouteMiddleware(async (to) => {
   const userRole = session.value?.user?.role
 
   const isAdminRoute = to.path.startsWith('/admin')
+  const isCoachRoute = to.path.startsWith('/coach')
   const isReaderRoute = to.path.startsWith('/reader')
 
-  if (!isLoggedIn && (isAdminRoute || isReaderRoute)) {
+  if (!isLoggedIn && (isAdminRoute || isCoachRoute || isReaderRoute)) {
     return navigateTo('/auth')
   }
 
@@ -18,20 +19,22 @@ export default defineNuxtRouteMiddleware(async (to) => {
   }
 
   if (to.path === '/auth') {
-    const requestedRole = Array.isArray(to.query.role) ? to.query.role[0] : to.query.role
-
-    if (requestedRole === 'admin') {
-      return navigateTo(userRole === 'admin' ? '/admin' : '/reader/profile')
+    if (userRole === 'admin') {
+      return navigateTo('/admin')
     }
 
-    if (requestedRole === 'reader') {
-      return navigateTo('/reader/profile')
+    if (userRole === 'coach') {
+      return navigateTo('/coach')
     }
 
-    return navigateTo(userRole === 'admin' ? '/admin' : '/reader/profile')
+    return navigateTo('/reader/profile')
   }
 
   if (isAdminRoute && userRole !== 'admin') {
+    return navigateTo(userRole === 'coach' ? '/coach' : '/reader/profile')
+  }
+
+  if (isCoachRoute && userRole !== 'admin' && userRole !== 'coach') {
     return navigateTo('/reader/profile')
   }
 })
