+++ app/pages/coach/styles/index.css
@@ -1,4 +1,4 @@
-.admin-home {
+.coach-home {
   max-width: 1100px;
   margin: 0 auto;
   padding: 2rem;
