+++ server/api/announcement/index.ts
@@ -1,6 +1,6 @@
 import { prisma } from '../../utils/prisma'
 import { getQuery, createError } from 'h3'
-import { requireAdmin, requireSession } from '../../utils/require-session'
+import { requireCoach, requireSession } from '../../utils/require-session'
 import { announcementCreateSchema } from '../../utils/schemas'
 
 // GET /api/announcement?active=true to get only active announcements
@@ -25,7 +25,7 @@ export default defineEventHandler(async (event) => {
 
   if (method === 'POST') {
 
-    const session = await requireAdmin(event)
+    const session = await requireCoach(event)
 
     const body = announcementCreateSchema.safeParse(await readBody(event))
