+++ server/utils/require-session.ts
@@ -27,3 +27,16 @@ export async function requireAdmin(event: H3Event) {
 
   return session
 }
+
+export async function requireCoach(event: H3Event) {
+  const session = await requireSession(event)
+
+  if (session.user.role !== 'coach' && session.user.role !== 'admin') {
+    throw createError({
+      statusCode: 403,
+      statusMessage: 'Forbidden',
+    })
+  }
+
+  return session
+}
